+++ NEWS.md
@@ -1,6 +1,7 @@
 # openair 2.9-99
 
 - add Daily Air Quality Index (DAQI) to `importAURN` as `data_type`.
+- add rolling 8 and 24 hour statistics to `importAURN` family as `data_type`.
 
 # openair 2.9-1